+++ tests/unit_tests/agents/deer_flow/test_server_manager.py
@@ -8,7 +8,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
-from genai_tk.extra.agents.deer_flow.server_manager import DeerFlowServerManager
+from genai_tk.agents.deer_flow.server_manager import DeerFlowServerManager
 
 # ---------------------------------------------------------------------------
 # Fixtures
@@ -73,7 +73,7 @@ async def test_start_skips_popen_when_already_running(tmp_path: Path) -> None:
     mgr = _manager(tmp_path)
     with (
         patch.object(mgr, "_check_url", new_callable=AsyncMock, return_value=True),
-        patch("genai_tk.extra.agents.deer_flow.server_manager.subprocess.Popen") as mock_popen,
+        patch("genai_tk.agents.deer_flow.server_manager.subprocess.Popen") as mock_popen,
     ):
         await mgr.start()
         mock_popen.assert_not_called()
@@ -94,7 +94,7 @@ async def test_start_launches_subprocesses(tmp_path: Path) -> None:
 
     with (
         patch.object(mgr, "_check_url", new_callable=AsyncMock, return_value=False),
-        patch("genai_tk.extra.agents.deer_flow.server_manager.subprocess.Popen", return_value=mock_proc) as mock_popen,
+        patch("genai_tk.agents.deer_flow.server_manager.subprocess.Popen", return_value=mock_proc) as mock_popen,
         patch.object(mgr, "_wait_for_ready", new_callable=AsyncMock),
     ):
         await mgr.start()
@@ -118,7 +118,7 @@ async def test_start_langgraph_command_contains_dev(tmp_path: Path) -> None:
 
     with (
         patch.object(mgr, "_check_url", new_callable=AsyncMock, return_value=False),
-        patch("genai_tk.extra.agents.deer_flow.server_manager.subprocess.Popen", side_effect=_popen),
+        patch("genai_tk.agents.deer_flow.server_manager.subprocess.Popen", side_effect=_popen),
         patch.object(mgr, "_wait_for_ready", new_callable=AsyncMock),
     ):
         await mgr.start()
@@ -144,7 +144,7 @@ async def test_start_gateway_command_contains_uvicorn(tmp_path: Path) -> None:
 
     with (
         patch.object(mgr, "_check_url", new_callable=AsyncMock, return_value=False),
-        patch("genai_tk.extra.agents.deer_flow.server_manager.subprocess.Popen", side_effect=_popen),
+        patch("genai_tk.agents.deer_flow.server_manager.subprocess.Popen", side_effect=_popen),
         patch.object(mgr, "_wait_for_ready", new_callable=AsyncMock),
     ):
         await mgr.start()
@@ -173,8 +173,8 @@ async def test_stop_sends_sigterm_when_owns_servers(tmp_path: Path) -> None:
     mgr._gw_proc = mock_proc
 
     with (
-        patch("genai_tk.extra.agents.deer_flow.server_manager.os.killpg") as mock_killpg,
-        patch("genai_tk.extra.agents.deer_flow.server_manager.os.getpgid", return_value=1234),
+        patch("genai_tk.agents.deer_flow.server_manager.os.killpg") as mock_killpg,
+        patch("genai_tk.agents.deer_flow.server_manager.os.getpgid", return_value=1234),
     ):
         await mgr.stop()
         assert mock_killpg.call_count == 2
@@ -190,7 +190,7 @@ async def test_stop_noop_when_not_owned(tmp_path: Path) -> None:
     mgr = _manager(tmp_path)
     mgr._owns_servers = False
 
-    with patch("genai_tk.extra.agents.deer_flow.server_manager.os.killpg") as mock_killpg:
+    with patch("genai_tk.agents.deer_flow.server_manager.os.killpg") as mock_killpg:
         await mgr.stop()
         mock_killpg.assert_not_called()
