+++ tests/unit_tests/agents/deer_flow/test_validation.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from genai_tk.extra.agents.deer_flow.profile import (
+from genai_tk.agents.deer_flow.profile import (
     DeerFlowProfile,
     InvalidModeError,
     MCPServerNotFoundError,
